refactor
Move sign-up functionality to auth controller
